third_party/blink/web_tests/inspector-protocol/css/cssom-matching-rules-multiple-expected.txt

The test verifies CSS.getMatchedStylesForNode when used concurrently with multiple CSSOM modifications.
Running test: testModifyRule
--------------
Original rule:
*#test* {    regular
    color: red; @[2:4-2:15]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[6:4-6:17]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[10:4-10:16]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    width: 10%; @[14:4-14:15]
    width: 10%; @[undefined-undefined]
}
*#test* {    regular
    width: 20%; @[18:4-18:15]
    width: 20%; @[undefined-undefined]
}
*#test* {    regular
    width: 30%; @[22:4-22:15]
    width: 30%; @[undefined-undefined]
}
*#test* {    regular
    width: 40%; @[26:4-26:15]
    width: 40%; @[undefined-undefined]
}
*#test* {    regular
    width: 50%; @[30:4-30:15]
    width: 50%; @[undefined-undefined]
}
*#test* {    regular
    width: 60%; @[34:4-34:15]
    width: 60%; @[undefined-undefined]
}
Mutating 3rd:
---------------
*#test* {    regular
    color: red; @[2:4-2:15]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[6:4-6:17]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[10:4-10:16]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    color: red; @[11:9-11:20]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    width: 20%; @[18:4-18:15]
    width: 20%; @[undefined-undefined]
}
*#test* {    regular
    width: 30%; @[22:4-22:15]
    width: 30%; @[undefined-undefined]
}
*#test* {    regular
    width: 40%; @[26:4-26:15]
    width: 40%; @[undefined-undefined]
}
*#test* {    regular
    width: 50%; @[30:4-30:15]
    width: 50%; @[undefined-undefined]
}
*#test* {    regular
    width: 60%; @[34:4-34:15]
    width: 60%; @[undefined-undefined]
}
Mutating 4th:
--------------
*#test* {    regular
    color: red; @[2:4-2:15]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[6:4-6:17]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[10:4-10:16]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    color: red; @[11:9-11:20]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[11:30-11:43]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    width: 30%; @[22:4-22:15]
    width: 30%; @[undefined-undefined]
}
*#test* {    regular
    width: 40%; @[26:4-26:15]
    width: 40%; @[undefined-undefined]
}
*#test* {    regular
    width: 50%; @[30:4-30:15]
    width: 50%; @[undefined-undefined]
}
*#test* {    regular
    width: 60%; @[34:4-34:15]
    width: 60%; @[undefined-undefined]
}
Mutating 5th:
--------------
*#test* {    regular
    color: red; @[2:4-2:15]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[6:4-6:17]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[10:4-10:16]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    color: red; @[11:9-11:20]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[11:30-11:43]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[11:53-11:65]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    width: 40%; @[26:4-26:15]
    width: 40%; @[undefined-undefined]
}
*#test* {    regular
    width: 50%; @[30:4-30:15]
    width: 50%; @[undefined-undefined]
}
*#test* {    regular
    width: 60%; @[34:4-34:15]
    width: 60%; @[undefined-undefined]
}
Delete first 3:
---------------
*#test* {    regular
    color: red; @[2:4-2:15]
    color: red; @[undefined-undefined]
}
*#test* {    regular
    color: green; @[6:4-6:17]
    color: green; @[undefined-undefined]
}
*#test* {    regular
    color: blue; @[10:4-10:16]
    color: blue; @[undefined-undefined]
}
*#test* {    regular
    width: 40%; @[26:4-26:15]
    width: 40%; @[undefined-undefined]
}
*#test* {    regular
    width: 50%; @[30:4-30:15]
    width: 50%; @[undefined-undefined]
}
*#test* {    regular
    width: 60%; @[34:4-34:15]
    width: 60%; @[undefined-undefined]
}
